+++ db19/big_test.go
@@ -155,5 +155,5 @@ func createData(db *Database, tables []string, i, n int) {
 }
 
 func fromHash(table string, max int) int {
-	return int(1 + hash.String(table)%uint32(max))
+	return int(1 + hash.String(table)%uint64(max))
 }
